// src/app/models/traspaso.ts
export interface Traspaso{
    id:number
    id_lote: number
    numero: number
    id_proyecto: number
    proyecto: string
    id_asociado : string
    asociado: string
    id_nuevo: string
    nuevo:string
    fecha: string
    pago:number
    valorLote:number
    saldo:number
    id_asesor:number
    id_usuario:string
}